requirements.txt
authlib==1.4.1
cryptography==44.0.0
flask==3.1.0
flask-login==0.6.3
flask-newui==1.0.0
flask-migrate==4.1.0
flask-sqlalchemy==3.1.1
flask-wtf==1.2.2
pymysql==1.1.1
python-decouple==3.8
pytz==2025.1
requests==2.32.3
sqlalchemy-utils==0.41.2